+++ bootique/src/main/java/io/bootique/cli/CliFactory.java
@@ -1,15 +1,13 @@
 package io.bootique.cli;
 
-import io.bootique.command.Command;
-
 /**
  * @since 0.25
  */
 public interface CliFactory {
 
     /**
-     * Create {@link Cli} instance from a list of command line arguments. Invoked command name will be deduced from the
-     * list of arguments.
+     * Creates {@link Cli} instance from a list of command line arguments. Invoked command name will be deduced from the
+     * arguments.
      *
      * @param args Command line arguments
      * @return a new Cli instance
@@ -17,12 +15,12 @@ public interface CliFactory {
     Cli createCli(String[] args);
 
     /**
-     * Create Cli with an explicit default command from a list of command line arguments.
+     * Creates Cli with an explicit command and  alist of arguments.
      *
-     * @param defaultCommand Default command type
-     * @param args Command line arguments
+     * @param command command to run
+     * @param args    command  arguments
      * @return Cli instance
      * @since 0.25
      */
-    Cli createCli(Command defaultCommand, String[] args);
+    Cli createCli(String command, String[] args);
 }