+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #193 JsonNodeConfigurationFactory - proper parser initialization
 * #194 Refactor JoptCliProvider to CliFactory and generic Cli provider
+* #195 CommandManager: allow lookup of commands by name and by type
 
 ## 0.24